app/config.py
import os
from pathlib import Path
from typing import List, Optional
from pydantic_settings import BaseSettings
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

class Settings(BaseSettings):
    # Core Application Settings
    app_name: str = "MahaVistaar AI API"
    environment: str = os.getenv("ENVIRONMENT", "production")
    debug: bool = False
    base_dir: Path = Path(__file__).resolve().parent.parent
    secret_key: str = os.getenv("SECRET_KEY", "your-secret-key-change-in-production")
    timezone: str = "Asia/Kolkata"

    # Server Configuration
    host: str = "0.0.0.0"
    port: int = 8000
    api_prefix: str = "/api"
    rate_limit_requests_per_minute: int = 1000

    # Security Settings
    allowed_origins: List[str] = os.getenv("ALLOWED_ORIGINS", "*").split(",")
    allowed_credentials: bool = True
    allowed_methods: List[str] = ["*"]
    allowed_headers: List[str] = ["*"]

    # JWT Configuration
    jwt_algorithm: str = "RS256"
    jwt_public_key_path: str = os.getenv("JWT_PUBLIC_KEY_PATH", "jwt_public_key.pem")
    jwt_private_key_path: Optional[str] = os.getenv("JWT_PRIVATE_KEY_PATH")

    # Worker Settings
    uvicorn_workers: int = os.cpu_count() or 1

    # Redis Settings
    redis_host: str = "localhost"
    redis_port: int = 6379
    redis_db: int = 0
    redis_key_prefix: str = "sva-cache-"
    redis_socket_connect_timeout: int = 10
    redis_socket_timeout: int = 10
    redis_max_connections: int = 100
    redis_retry_on_timeout: bool = True

    # Cache Configuration
    default_cache_ttl: int = 60 * 60 * 24  # 24 hours
    suggestions_cache_ttl: int = 60 * 30    # 30 minutes

    # PostgreSQL Database Configuration
    database_url: str = os.getenv("DATABASE_URL", "postgresql://postgres:postgres@localhost:5432/load_agri")
    db_pool_size: int = int(os.getenv("DB_POOL_SIZE", "20"))
    db_max_overflow: int = int(os.getenv("DB_MAX_OVERFLOW", "10"))

    # NMIS Data Scraper Configuration
    nmis_api_base_url: str = "http://nmis.et/api"
    scraper_timeout: int = 30  # seconds
    scraper_enabled: bool = os.getenv("SCRAPER_ENABLED", "true").lower() == "true"

    # Logging Configuration
    log_level: str = "INFO"
    log_format: str = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"

    # External Service URLs
    telemetry_api_url: str = "https://vistaar.kenpath.ai/observability-service/action/data/v3/telemetry"
    bhashini_api_url: str = ""
    ollama_endpoint_url: Optional[str] = None
    marqo_endpoint_url: Optional[str] = None
    inference_endpoint_url: Optional[str] = None

    # External Service API Keys
    openai_api_key: Optional[str] = None
    sarvam_api_key: Optional[str] = None
    meity_api_key_value: Optional[str] = None
    logfire_token: Optional[str] = None
    bhashini_api_key: str = ""
    eleven_labs_api_key: str = ""
    inference_api_key: Optional[str] = None
    gemini_api_key: Optional[str] = None
    mapbox_api_token: Optional[str] = None

    # AWS Configuration
    aws_access_key_id: Optional[str] = None
    aws_secret_access_key: Optional[str] = None
    aws_region: Optional[str] = None
    aws_s3_bucket: Optional[str] = None

    # LLM Configuration
    llm_provider: Optional[str] = None
    llm_model_name: Optional[str] = None
    marqo_index_name: Optional[str] = None

    # Azure Configuration
    azure_foundary_api_key: Optional[str] = None
    azure_foundary_region: Optional[str] = None

    # RAG Configuration
    rag_provider: str = os.getenv("RAG_PROVIDER", "marqo")  # "marqo" or "cosdata"

    # Cosdata Configuration
    cosdata_endpoint_url: Optional[str] = os.getenv("COSDATA_ENDPOINT_URL")
    cosdata_api_key: Optional[str] = os.getenv("COSDATA_API_KEY")
    cosdata_collection_name: str = os.getenv("COSDATA_COLLECTION_NAME", "oan-collection")

    # Embedding Configuration (for Cosdata - uses sentence-transformers locally)
    embedding_model_name: str = os.getenv("EMBEDDING_MODEL_NAME", "intfloat/multilingual-e5-large")

    # OpenTelemetry / SigNoz Configuration
    otel_service_name: str = os.getenv("OTEL_SERVICE_NAME", "oan-ai-api")
    otel_exporter_otlp_endpoint: str = os.getenv("OTEL_EXPORTER_OTLP_ENDPOINT", "http://localhost:4318")
    otel_enabled: bool = os.getenv("OTEL_ENABLED", "true").lower() == "true"

    class Config:
        env_file = ".env"
        extra = 'ignore'  # Ignore extra fields from .env

settings = Settings()
